refactor: Decompose DreamService into modular services (#10013)
